Change 3 functions to be #defines of others
I made these separate functions because I thought would make faster
code, but I realized that modern compilers should be able to optimize
the more general functions into the same code as the ones removed, given
that the parameters are known to be 0 at compile time and can easily
determined to always be false.

It's easier to maintain one version of a function than two, so this
commit favors that.