Add is_utf8_non_invariant_string()
This function tells whether or not its argument is legal UTF-8 that
isn't invariant under UTF-8 or not.

This paradigm is used in several places in the perl core to decide
whether to turn on an SV's utf8 flag.  None of those places realized
that there was a simple way to avoid rescanning the string (though
perhaps an C optimizer would).  This commit uses that method; the next
commits will convert to use this function.