XXX msg: Use is_utf8_invariant_string() more
Now that this function was changed to do word-at-a time searching in
commit XXX,
we can more quickly find the first variant byte in a string, if any.
Given that a lot of usage of Perl is on ASCII data, it makes sense to
try this first before any byte-at-a-time processing.

Since Perl can be used on things that are mostly non-ASCII, we give up
at the first such one, and process the rest of the string byte-by-byte.
Otherwise we could have a pipeline of finding the next variant quickly,
but this would only be faster if variants were rare, which I don't feel
we can be confident about, after finding at least one.