Emulate some C99 macros and typedefs
On systems that don't have them, this emulates

u?intmax_t
U?INT(16|32)_C
U?INTMAX_C

In addition, it moves the pre-existing emulation of U?INT64_C from
handy.h to perl.h.  This is because there was duplicate-ish logic in the
two files, and the handy.h version appears to be better thought out.

It converts the couple of places in core that were using the other
macros to use the C99 names.